refactor: extract repo loading logic to object methods
